refactored imports and error classes
